- Couple minor changes
git-svn-id: svn://70.113.44.214/VanaMirror/trunk@2722 09c1253f-3d09-8e40-9719-4320e56601b5

+++ src/ChannelServer/Mob.h
@@ -134,7 +134,7 @@ public:
 	Mob * getSponge() const { return sponge; }
 	const MobInfo * getInfo() const { return info; }
 	uint8_t getSkillCount() const { return info->skillcount; }
-	size_t getSpawnCount() const { return spawns.size(); }
+	int16_t getSpawnCount() const { return static_cast<int16_t>(spawns.size()); }
 	unordered_map<int32_t, Mob *> getSpawns() const { return spawns; }
 
 	Timer::Container * getTimers() const { return timers.get(); }
